Fix phpunit tests.
